+++ R/fortify_docx.R
@@ -236,6 +236,9 @@ docx_summary <- function(x, preserve = FALSE, remove_fields = FALSE, detailed = 
   if (remove_fields) {
     instrText_nodes <- xml_find_all(x$doc_obj$get(), "//w:instrText")
     xml_remove(instrText_nodes)
+    
+    fldData_nodes <- xml_find_all(x$doc_obj$get(), "//w:fldData")
+    xml_remove(fldData_nodes)
   }
 
   all_nodes <- xml_find_all(x$doc_obj$get(), "/w:document/w:body/*[self::w:p or self::w:tbl]")